Changed system to use AVR-level UART communication.  I just wasn't sure the CEENbot API would provide the proper functionality that I expected. Changed the definitions so that they could be incremented to the next command.
The state changes are now induced by a "Receive flag", or rxflag.  When this is high, it implies that microcontroller has received a byte in the UART register and stored it into cmd, a volatile char.

During each state, I have the LCD simply print out what it is doing, mainly for debugging purposes.  Whenever a turn is executed, the bot will resume whatever it was previously doing; if the bot was going forward, it would execute its turn and resume going forward without further instruction.

The turns have now been issued in a wait state, simply due to the fact that a complete 90 degree turn is expected and will not be interrupted before it is complete.  For turning around, the bot will simply turn right for 300 turns, or 180 degrees.